feat(automation): gate memory deep-link dispatch behind __E2E__
Moves the memory-profile deep-link handler body out of
src/hooks/useDeepLinking.ts into src/__automation__/deepLink.ts and
wraps the call in a runtime __E2E__ gate with a deferred require()
inside the branch.

When __E2E__ inlines to false at build time, the whole branch is
dead-code-eliminated and src/__automation__/deepLink.ts is never
pulled into the prod bundle. The chat-pal deep link and its
useCallback dep array remain unchanged (non-E2E surface).

dispatchAutomationDeepLink returns a boolean so the caller can fall
through to other handlers if the host is not 'memory' — matches the
existing early-return pattern in useDeepLinking.
