+++ src/__automation__/index.ts
@@ -1,1 +1,2 @@
 export {AutomationBridge} from './AutomationBridge';
+export {dispatchAutomationDeepLink} from './deepLink';
